+++ packages/medusa-core-utils/src/validator.js
@@ -17,4 +17,73 @@ Joi.address = () => {
   })
 }
 
+Joi.dateFilter = () => {
+  return Joi.object({
+    lt: Joi.alternatives(Joi.date().timestamp("unix"), Joi.date()),
+    gt: Joi.alternatives(Joi.date().timestamp("unix"), Joi.date()),
+    gte: Joi.alternatives(Joi.date().timestamp("unix"), Joi.date()),
+    lte: Joi.alternatives(Joi.date().timestamp("unix"), Joi.date()),
+  })
+}
+
+Joi.orderFilter = () => {
+  return Joi.object().keys({
+    id: Joi.string(),
+    q: Joi.string(),
+    status: Joi.array()
+      .items(
+        Joi.string().valid(
+          "pending",
+          "completed",
+          "archived",
+          "canceled",
+          "requires_action"
+        )
+      )
+      .single(),
+    fulfillment_status: Joi.array()
+      .items(
+        Joi.string().valid(
+          "not_fulfilled",
+          "fulfilled",
+          "partially_fulfilled",
+          "shipped",
+          "partially_shipped",
+          "canceled",
+          "returned",
+          "partially_returned",
+          "requires_action"
+        )
+      )
+      .single(),
+    payment_status: Joi.array()
+      .items(
+        Joi.string().valid(
+          "captured",
+          "awaiting",
+          "not_paid",
+          "refunded",
+          "partially_refunded",
+          "canceled",
+          "requires_action"
+        )
+      )
+      .single(),
+    display_id: Joi.string(),
+    cart_id: Joi.string(),
+    offset: Joi.string(),
+    limit: Joi.string(),
+    expand: Joi.string(),
+    fields: Joi.string(),
+    customer_id: Joi.string(),
+    email: Joi.string(),
+    region_id: Joi.string(),
+    currency_code: Joi.string(),
+    tax_rate: Joi.string(),
+    canceled_at: Joi.dateFilter(),
+    created_at: Joi.dateFilter(),
+    updated_at: Joi.dateFilter(),
+  })
+}
+
 export default Joi